# api-spec/src/main/resources/openapi/discussion.yaml
openapi: 3.0.0
info:
  title: Študentské fórum FSA
  version: 1.0.0
  description: API pre správu diskusných príspevkov a úloh k lekciám.
paths:
...
  /courses/{courseId}/topics:
    get:
      summary: Get topics of a course forum
      operationId: getTopicsByCourse
      tags: [ Forum ]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Paged list of topics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicsResponse'

    post:
      summary: Create new topic in a course forum
      operationId: createTopic
      tags: [ Forum ]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTopicRequest'
      responses:
        '201':
          description: Topic created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Topic'
  /topics/{topicId}:
    get:
      summary: Get a topic by ID
      operationId: getTopicById
      tags: [ Forum ]
      parameters:
        - name: topicId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Topic details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Topic'
        '404': { description: Topic not found }

    put:
      summary: Update topic (title or status)
      operationId: updateTopic
      tags: [ Forum ]
      parameters:
        - name: topicId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTopicRequest'
      responses:
        '200':
          description: Topic updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Topic'
        '404': { description: Topic not found }

    delete:
      summary: Delete topic
      operationId: deleteTopic
      tags: [ Forum ]
      parameters:
        - name: topicId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: Topic deleted }
        '404': { description: Topic not found }
  /topics/{topicId}/posts:
    get:
      summary: Get root posts (questions) of topic
      operationId: getPostsByTopic
      tags: [ Forum ]
      parameters:
        - name: topicId
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Paged list of posts (only first-level)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostsResponse'
        '404': { description: Topic not found }

    post:
      summary: Create a root post (question) in topic
      operationId: createPost
      tags: [ Forum ]
      parameters:
        - name: topicId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404': { description: Topic not found }
  /posts/{postId}:
    get:
      summary: Get a single post with its replies (recursive tree)
      operationId: getPostById
      tags: [ Forum ]
      parameters:
        - name: postId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Post with nested replies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404': { description: Post not found }

    put:
      summary: Update a post
      operationId: updatePost
      tags: [ Forum ]
      parameters:
        - name: postId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        '200':
          description: Post updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404': { description: Post not found }

    delete:
      summary: Delete a post
      operationId: deletePost
      tags: [ Forum ]
      parameters:
        - name: postId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: Post deleted }
        '404': { description: Post not found }
  /posts/{postId}/replies:
    post:
      summary: Reply to a post
      operationId: createReply
      tags: [ Forum ]
      parameters:
        - name: postId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReplyRequest'
      responses:
        '201':
          description: Reply created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404': { description: Post not found }
...
  /courses:
    get:
      summary: Get all courses
      operationId: getAllCourses
      tags:
        - Course
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoursesResponse'
    post:
      summary: Create new course
      operationId: createCourse
      tags:
        - Course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseRequest'
      responses:
        '201':
          description: Course created
  /courses/{id}/lessons:
    get:
      summary: Get lessons by course ID
      operationId: getLessonsByCourseId
      tags:
        - Lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of lessons for the course
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lesson'
        '404':
          description: Course not found
...
  /users:
    get:
      summary: Get all users
      operationId: getAllUsers
      tags:
        - User
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
    post:
      summary: Create a new user
      operationId: createUser
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'   # ← тут ссылка!
      responses:
        '201':
          description: User created
  /users/{userId}/courses/{courseId}:
    post:
      summary: Enroll user to course
      operationId: enrollUserToCourse
      tags:
        - User
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User enrolled to course
        '404':
          description: User or Course not found
components:
  schemas:
    CreateTopicRequest:
      type: object
      required: [ title ]
      properties:
        title:
          type: string
        lessonId:
          type: integer
          format: int64
          nullable: true
          description: Optional lesson reference; null → general topic
    UpdateTopicRequest:
      allOf:
        - $ref: '#/components/schemas/CreateTopicRequest'
        - type: object
          required: [ id ]
          properties:
            id:
              type: integer
              format: int64
            status:
              $ref: '#/components/schemas/TopicStatus'
    Topic:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        courseId:
          type: integer
          format: int64
        lessonId:
          type: integer
          format: int64
          nullable: true
        status:
          $ref: '#/components/schemas/TopicStatus'
        createdAt:
          type: string
          format: date-time
        createdById:
          type: integer
          format: int64
        postsCount:
          type: integer
          format: int32
        lastActivityAt:
          type: string
          format: date-time
    TopicStatus:
      type: string
      enum: [ OPEN, CLOSED, ARCHIVED ]
    TopicsResponse:
      type: object
      properties:
        topics:
          type: array
          items: { $ref: '#/components/schemas/Topic' }
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    CreatePostRequest:
      type: object
      required: [ content ]
      properties:
        content:
          type: string
    CreateReplyRequest:
      type: object
      required: [ content ]
      properties:
        content:
          type: string
    UpdatePostRequest:
      type: object
      required: [ id, content ]
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
    Post:
      type: object
      properties:
        id:
          type: integer
          format: int64
        topicId:
          type: integer
          format: int64
        parentId:
          type: integer
          format: int64
          nullable: true
        authorName:
          type: string
          description: "Human-readable name of the post's author"
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/PostStatus'
        replies:
          type: array
          description: Recursive list of replies; omitted in collections
          items: { $ref: '#/components/schemas/Post' }
    PostStatus:
      type: string
      enum: [ ACTIVE, HIDDEN, DELETED ]
    PostsResponse:
      type: object
      properties:
        posts:
          type: array
          items: { $ref: '#/components/schemas/Post' }
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
...
    TestCase:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: "Auto-generated test case ID"
        input:
          type: string
          description: "The input data for this test case"
        expectedOutput:
          type: string
          description: "The expected result when running the assignment code"
    CreateCourseRequest:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
        description:
          type: string
...
    Assignment:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: "Unique identifier of the assignment"
        description:
          type: string
        teacherCode:
          type: string
        templateCode:
          type: string
        expectedOutput:
          type: string
        outputType:
          $ref: '#/components/schemas/AssignmentOutputType'
        testCases:
          type: array
          items:
            $ref: '#/components/schemas/TestCase'
          description: "List of all test cases associated with this assignment"
    AssignmentOutputType:
      type: string
      enum:
        - STRING
        - INT
        - JSON
      description: "Type of the expected output used for automatic checking"
...
    Question:
      type: object
      properties:
        id:
          type: integer
          format: int64
        text:
          type: string
        type:
          $ref: '#/components/schemas/QuestionType'
        options:
          type: array
          items:
            $ref: '#/components/schemas/AnswerOption'
...
    QuestionType:
      type: string
      enum:
        - SINGLE_CHOICE
        - MULTIPLE_CHOICE
        - OPEN
      description: Тип вопроса
...
    AnswerOption:
      type: object
      properties:
        id:
          type: integer
          format: int64
        text:
          type: string
        correct:
          type: boolean

    # ——— Обновление Lesson, чтобы отдать вопросы вместе с заданиями ———
    Lesson:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        courseId:
          type: integer
          format: int64
        assignments:
          type: array
          items:
            $ref: '#/components/schemas/Assignment'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
    Course:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        lessons:
          type: array
          items:
            $ref: '#/components/schemas/Lesson'
    CreateUserRequest:                                 # ← вот он
      type: object
      required:
        - givingName
        - familyName
        - email
        - role
      properties:
        givingName:
          type: string
        familyName:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - Student
            - Teacher
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        givingName:
          type: string
        familyName:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - Student
            - Teacher
        rating:
          type: integer
          format: int32
          description: Рейтинг пользователя
        courses:
          type: array
          items:
            $ref: '#/components/schemas/Course'
    CoursesResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: '#/components/schemas/Course'
    UsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'